cmake files: Asset module doesn't depend on anything (since it's an umbrella module)

+++ code/Modules/Sound/Core/SoundEffectSetup.cc
@@ -8,7 +8,7 @@ namespace Oryol {
 
 //------------------------------------------------------------------------------
 SoundEffectSetup
-SoundEffectSetup::FromSampleFunc(int32 duration, SampleFuncT sampleFunc) {
+SoundEffectSetup::FromSampleFunc(float32 duration, SampleFuncT sampleFunc) {
     SoundEffectSetup setup;
     setup.Duration = duration;
     setup.SampleFunc = sampleFunc;
